# docs/volue_swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Volue Algo Trader Power API",
    "description": "Your go-to API for automated power trading",
    "contact": {
      "name": "Volue Trading",
      "url": "https://volue.com",
      "email": "support.trading@volue.com"
    },
    "version": "2.0"
  },
  "paths": {
...
    "/v2/order": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Gets orders with specified filter values",
        "operationId": "Order_GET_v2/order",
        "parameters": [
          {
            "name": "states",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderFilterState"
              }
            }
          },
          {
            "name": "source",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderSource"
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Exchange"
            }
          },
          {
            "name": "portfolios",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "texts",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "contractStarts",
            "in": "query",
            "description": "Returns strategies where the contract's delivery begins only at the specified time.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          {
            "name": "contractTypes",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ContractType"
              }
            }
          },
          {
            "name": "deliveryAreas",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DeliveryArea"
              }
            }
          },
          {
            "name": "deliveryDays",
            "in": "query",
            "description": " (Deprecated: DeliveryDays is no longer used. Use DeliveryStartTime and DeliveryEndTime instead.)",
            "deprecated": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "date"
              },
              "deprecated": true
            }
          },
          {
            "name": "deliveryStartTime",
            "in": "query",
            "description": "Returns strategies where the delivery starts at the specified time or any time afterward.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "deliveryEndTime",
            "in": "query",
            "description": "Defines the latest allowable delivery end time, limiting the visible time frame.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Creates orders with specified values",
        "operationId": "Order_POST_v2/order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrdersModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrdersModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderActionStatusModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "Updates existing orders with specified values",
        "operationId": "Order_PUT_v2/order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyOrdersModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyOrdersModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderActionStatusModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Deletes orders with specified IDs",
        "operationId": "Order_DELETE_v2/order",
        "parameters": [
          {
            "name": "orderIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteOrderRangeFilterModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteOrderRangeFilterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderActionStatusModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/v2/order/{orderId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Gets an order with the specified ID",
        "operationId": "Order_GET_v2/order/{orderId}",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "Order"
        ],
        "summary": "Checks if an order with a specified ID exists",
        "operationId": "Order_HEAD_v2/order/{orderId}",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/reporting/orders": {
      "get": {
        "tags": [
          "Reporting"
        ],
        "summary": "Gets historical manual orders executed between specified delivery days",
        "description": "Only non-active manual orders are retrieved",
        "operationId": "Reporting_GET_v2/reporting/orders",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Exchange"
            }
          },
          {
            "name": "deliveryAreas",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DeliveryArea"
              }
            }
          },
          {
            "name": "portfolios",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "The start delivery day",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "The end delivery day",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContractModel": {
        "type": "object",
        "properties": {
          "deliveryStartTime": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryEndTime": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/ContractType"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ContractRequestModel": {
        "type": "object",
        "properties": {
          "deliveryStartTime": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryEndTime": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/ContractType"
          }
        },
        "additionalProperties": false
      },
      "ContractType": {
        "enum": [
          "Hour",
          "HalfHour",
          "QuarterHour",
          "Block",
          "Block2H",
          "Block4H",
          "Block3Plus4",
          "BlockBase",
          "BlockPeak",
          "BlockExtendedPeak",
          "BlockOvernight",
          "CustomBlock"
        ],
        "type": "string"
      },
      "CreateOrderModel": {
        "required": [
          "contract",
          "deliveryArea",
          "exchange",
          "orderId",
          "price",
          "quantity",
          "side",
          "type"
        ],
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string",
            "nullable": true
          },
          "balancingGroup": {
            "type": "string",
            "nullable": true
          },
          "contract": {
            "$ref": "#/components/schemas/ContractRequestModel"
          },
          "deliveryArea": {
            "$ref": "#/components/schemas/DeliveryArea"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "groupRestriction": {
            "$ref": "#/components/schemas/OrderGroupRestriction"
          },
          "isAutoSegmentTransfer": {
            "type": "boolean",
            "nullable": true,
            "deprecated": true
          },
          "peakQuantity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "portfolio": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          },
          "targetDeliveryArea": {
            "$ref": "#/components/schemas/DeliveryArea"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "validityEndTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CreateOrdersModel": {
        "required": [
          "orders"
        ],
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderModel"
            }
          },
          "createAsHibernated": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "DeleteOrderRangeFilterModel": {
        "type": "object",
        "properties": {
          "orderIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryArea": {
        "enum": [
          "Unspecified",
          "APG",
          "TNG",
          "TTG",
          "RTE",
          "AMP",
          "SGD",
          "_50HZT",
          "TTN",
          "ELIA",
          "DK1",
          "DK1A",
          "DK2",
          "EE",
          "FI",
          "FRE",
          "HR",
          "LT",
          "LV",
          "NO1",
          "NO1A",
          "NO2",
          "NO2A",
          "NO3",
          "NO4",
          "NO5",
          "SE1",
          "SE2",
          "SE3",
          "SE4",
          "NGET",
          "NI",
          "ROI",
          "GB2",
          "GB1A",
          "ITAUST",
          "ITCALA",
          "ITCNOR",
          "ITCOAC",
          "ITCODC",
          "ITCPL",
          "ITCSUD",
          "ITFRAN",
          "ITGREC",
          "ITMALT",
          "ITMONT",
          "ITNORD",
          "ITSARD",
          "ITSICI",
          "ITSLOV",
          "ITSUD",
          "ITSVIZ",
          "REN",
          "REE",
          "ONE",
          "BG",
          "PL",
          "GE",
          "CEPS",
          "MAVIR",
          "TEL",
          "ELES",
          "PLA",
          "PLC",
          "LRI",
          "LRE",
          "ELE",
          "ELI",
          "ERI",
          "LBI",
          "LBE",
          "ADMIE",
          "SEPS",
          "N2EX"
        ],
        "type": "string"
      },
...
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Exchange": {
        "enum": [
          "EPEX",
          "Nordpool",
          "HUPX",
          "InternalMarket",
          "OMIE",
          "Southpool",
          "GME"
        ],
        "type": "string"
      },
...
      "ModifyOrderModel": {
        "required": [
          "orderId"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "description": "Either Price or OpenQuantity must be provided",
            "format": "double",
            "nullable": true
          },
          "portfolio": {
            "type": "string",
            "nullable": true
          },
          "openQuantity": {
            "type": "number",
            "description": "Either Price or OpenQuantity must be provided",
            "format": "double",
            "nullable": true
          },
          "peakQuantity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "validityEndTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModifyOrdersModel": {
        "required": [
          "modifications"
        ],
        "type": "object",
        "properties": {
          "modifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModifyOrderModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "OrderActionStatus": {
        "enum": [
          "Succeeded",
          "Failed",
          "TimedOut"
        ],
        "type": "string"
      },
      "OrderActionStatusModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/OrderActionStatus"
          }
        },
        "additionalProperties": false
      },
      "OrderFilterState": {
        "enum": [
          "Active",
          "Hibernating",
          "Pending"
        ],
        "type": "string"
      },
      "OrderGroupRestriction": {
        "enum": [
          "Unspecified",
          "Linked",
          "Valid"
        ],
        "type": "string"
      },
      "OrderModel": {
        "type": "object",
        "properties": {
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "contract": {
            "$ref": "#/components/schemas/ContractModel"
          },
          "deliveryArea": {
            "$ref": "#/components/schemas/DeliveryArea"
          },
          "targetDeliveryArea": {
            "$ref": "#/components/schemas/DeliveryArea"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "portfolio": {
            "type": "string",
            "nullable": true
          },
          "balancingGroup": {
            "type": "string",
            "nullable": true
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "strategyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "accountType": {
            "type": "string",
            "nullable": true
          },
          "marketOrderId": {
            "type": "string",
            "nullable": true
          },
          "isAutoSegmentTransfer": {
            "type": "boolean",
            "nullable": true,
            "deprecated": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "totalQuantity": {
            "type": "number",
            "format": "double"
          },
          "openQuantity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "peakQuantity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/OrderSource"
          },
          "state": {
            "$ref": "#/components/schemas/OrderState"
          },
          "strategyName": {
            "type": "string",
            "nullable": true
          },
          "marketUserName": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "validityEndTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "validityRestriction": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OrderSource": {
        "enum": [
          "Manual",
          "Strategy"
        ],
        "type": "string"
      },
      "OrderState": {
        "enum": [
          "Unknown",
          "Created",
          "Pending",
          "Active",
          "Hibernating",
          "PartiallyFilled",
          "Filled",
          "Cancelled",
          "Rejected"
        ],
        "type": "string"
      },
...
      "OrderType": {
        "enum": [
          "Unspecified",
          "Limit",
          "ImmediateOrCancel",
          "FillOrKill",
          "AllOrNothing",
          "Iceberg",
          "OTC",
          "Auction"
        ],
        "type": "string"
      },
...
      "Side": {
        "enum": [
          "Buy",
          "Sell"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Enter your Bearer JWT token string below (without the 'Bearer ' prefix)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": [
        "vts.vatp.api.read",
        "vts.vatp.api.write"
      ]
    }
  ],
...
}